requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.5
filelock==3.12.4
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai==0.28.0
openai-whisper==20230918
packaging==23.2
regex==2023.8.8
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
whitenoise==6.5.0
yarl==1.9.2
